中間点の実装をCoordObjectに依存しなくなったのでCoordObject削除 余分な記述消したりfloat使ってたところをdoubleにしたり

+++ Metasia.Core/Objects/IMetaAudiable.cs
@@ -9,7 +9,7 @@ namespace Metasia.Core.Objects
 {
 	public interface IMetaAudiable
 	{
-		public float Volume { get; set; }
+		public double Volume { get; set; }
 		public void AudioExpresser(ref AudioExpresserArgs e,  int frame);
 	}
 }
